Adding a button to toggle mode
